+++ inc/lilyParser/index.ts
@@ -2,7 +2,7 @@
 import {termDictionary, getDurationSubdivider} from "./lilyTerms";
 import LilyDocument from "./lilyDocument";
 import LilyInterpreter, {MusicTrack} from "./lilyInterpreter";
-import {replaceSourceToken, WHOLE_DURATION_MAGNITUDE} from "./utils";
+export * from "./utils";
 
 import * as measures from "./measures";
 import {createPianoRhythm} from "./pianoRhythm";
@@ -14,7 +14,6 @@ const LilyTerms = termDictionary;
 
 
 export {
-	replaceSourceToken,
 	LilyDocument,
 	LilyInterpreter,
 	MusicTrack,
@@ -22,5 +21,4 @@ export {
 	getDurationSubdivider,
 	measures,
 	createPianoRhythm,
-	WHOLE_DURATION_MAGNITUDE,
 };