+++ inc/staffSvg/staffToken.ts
@@ -167,6 +167,12 @@ export default class StaffToken {
 	}
 
 
+	get topAtSystem (): boolean {
+		return this.is("OCTAVE A") || this.is("CHORD_TEXT") || this.is("REPEAT_SIGN SEGNO")
+			|| this.is("REPEAT_SIGN CODA") || this.is("TEMPO_NOTEHEAD");
+	}
+
+
 	get logicY (): number {
 		return this.y + this.logicOffsetY;
 	}